backend/requirements.txt
# Backend Requirements - Versions exactes testées
# Ce fichier contient les versions EXACTES qui fonctionnent ensemble sans conflit
# Utilisez ce fichier si vous rencontrez des problèmes avec requirements.txt

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.4
pydantic-ai==0.0.14
python-dotenv==1.0.1

# AI Models (versions compatibles avec pydantic-ai 0.0.14)
openai==1.57.4
anthropic==0.40.0

# Pinecone & Embeddings
pinecone-client==5.0.1
sentence-transformers==3.3.1

# Utils
python-multipart==0.0.18

# Dépendances supplémentaires (installées automatiquement par les packages ci-dessus)
# httpx>=0.27.2
# jiter>=0.8.2
# typing-extensions>=4.12.2
# httpcore>=1.0.7
# anyio>=4.6.2
# sniffio>=1.3.1
# certifi>=2024.12.14
# h11>=0.14.0
# idna>=3.10
